done exercise 4.21

+++ part4/bloglist-backend/controllers/blogs.js
@@ -49,8 +49,31 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(404).json({
+      error: 'Can not find blog',
+    });
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid ',
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    user.blogs = user.blogs
+      .filter((blogId) => blogId.toString() !== blog._id.toString());
+    await user.save();
+    return response.status(204).end();
+  }
+
+  return response.status(403).end();
 });
 
 module.exports = blogsRouter;
